refactor(bin): separate event allocator from scheduler
Then the only primitive provided by the top-level event loop is run!
