+++ bin/simexp/main.cpp
@@ -67,15 +67,15 @@ int main(int argc, char *argv[]) {
     auto snooper = s.dispatcher.install_filter<AccessSnooper<DRAM>>({sim.id(), Event::Type::MemoryAccess});
     snooper->_id = 3;
     i64 *ptr = &sim.icount;
-    auto ev = s.make_event<ClockEvent>();
+    auto ev = s.allocator.make_event<ClockEvent>();
     ev->base.type = Event::Type::Clock;
     ev->base.recurs = true;
     ev->base.source = sim.id();
-    s.schedule(ev->base.event_index, 0);
+    s.scheduler.schedule(ev->base.event_index, 0);
     s.run([ptr, maxi]() { return *ptr >= maxi; });
-    ic = sim.icount, cc = s.current_tick(), wc = sim.wcount;
-    fmt::println("Executed {}, allocated {} and freed {} events", s._counters.executed, s._counters.allocated,
-                 s._counters.freed);
+    ic = sim.icount, cc = s.scheduler.current_tick(), wc = sim.wcount;
+    /*fmt::println("Executed {}, allocated {} and freed {} events", s._counters.executed, s._counters.allocated,
+                 s._counters.freed);*/
     fmt::println("Access memory {} times", snooper->access_count);
   }
